ai-services-local/sdxl_service/requirements.txt
fastapi
uvicorn[standard]
pillow
torch
torchvision
diffusers
transformers
accelerate
safetensors
